requirements.txt
appdirs==1.4.4
asgiref==3.4.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.0
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
djangorestframework==3.12.4
filelock==3.0.12
idna==2.10
importlib-metadata==1.6.1
importlib-resources==2.0.1
pyTelegramBotAPI==3.7.6
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.3
virtualenv==20.0.23
zipp==3.1.0
